Move ops/_yaml.py to ops/_private/yaml.py
Per Mattermost discussion, want to avoid adding individual files to
the top-level ops directory to keep that clean, so add a single _private
directory and we can put this (and any other) private/internal stuff
there.
