+++ test/test_helpers.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import io
 import os
 import pathlib
 import subprocess
@@ -20,9 +19,6 @@ import shutil
 import tempfile
 import unittest
 
-import yaml
-
-from ops import _yaml
 from ops.framework import Framework
 from ops.model import Model, _ModelBackend
 from ops.charm import CharmMeta
@@ -134,31 +130,3 @@ class BaseTestCase(unittest.TestCase):
         meta = CharmMeta()
         model = Model(meta, backend)
         return model
-
-
-class YAMLTest:
-    pass
-
-
-class TestYAML(unittest.TestCase):
-    def test_safe_load(self):
-        d = _yaml.safe_load('foo: bar\nbaz: 123\n')
-        self.assertEqual(len(d), 2)
-        self.assertEqual(d['foo'], 'bar')
-        self.assertEqual(d['baz'], 123)
-
-        # Should error -- it's not safe to load an instance of a user-defined class
-        with self.assertRaises(yaml.YAMLError):
-            _yaml.safe_load('!!python/object:test.test_helpers.YAMLTest {}')
-
-    def test_safe_dump(self):
-        s = _yaml.safe_dump({'foo': 'bar', 'baz': 123})
-        self.assertEqual(s, 'baz: 123\nfoo: bar\n')
-
-        f = io.StringIO()
-        s = _yaml.safe_dump({'foo': 'bar', 'baz': 123}, stream=f)
-        self.assertEqual(f.getvalue(), 'baz: 123\nfoo: bar\n')
-
-        # Should error -- it's not safe to dump an instance of a user-defined class
-        with self.assertRaises(yaml.YAMLError):
-            _yaml.safe_dump(YAMLTest())
